Fixed the folder structure.
