+++ UI/XPlat/flutter_template/android/app/google-services.json
@@ -1,1 +1,1 @@
-’u‚«Š·‚¦‚éB
+ç½®ãæ›ãˆã‚‹ã€‚
