+++ UI/XPlat/flutter_template/lib/components/importer.dart
@@ -1,0 +1,15 @@
+// dart
+import 'dart:async';
+// package
+import 'package:flutter/material.dart';
+
+// ...
+export 'package:flutter_template/configs/app_fcm.dart';
+export 'package:flutter_template/models/message_arguments.dart';
+
+export 'package:flutter_template/common/permissions.dart';
+export 'package:flutter_template/common/token_monitor.dart';
+export 'package:flutter_template/common/message_list.dart';
+export 'package:flutter_template/common/meta_card.dart';
+
+export 'package:flutter_template/components/message_view.dart';
