+++ UI/XPlat/flutter_template/lib/components/message_view.dart
@@ -1,18 +1,8 @@
-import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
+import 'package:firebase_messaging/firebase_messaging.dart';
 
-/// Message route arguments.
-class MessageArguments {
-  /// The RemoteMessage
-  final RemoteMessage message;
-
-  /// Whether this message caused the application to open.
-  final bool openedApplication;
-
-  // ignore: public_member_api_docs
-  MessageArguments(this.message, this.openedApplication)
-      : assert(message != null);
-}
+// ...
+import 'package:flutter_template/models/message_arguments.dart';
 
 /// Displays information about a [RemoteMessage].
 class MessageView extends StatelessWidget {
@@ -30,31 +20,31 @@ class MessageView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     MessageArguments? args =
-      ModalRoute.of(context)?.settings.arguments as MessageArguments;
+    ModalRoute.of(context)?.settings.arguments as MessageArguments;
 
-    RemoteMessage? message = args.message;
+    RemoteMessage? message = args?.message;
     RemoteNotification? notification = message?.notification;
 
     return Scaffold(
       appBar: AppBar(
-        title: Text(message.messageId ?? ""),
+        title: Text(message?.messageId ?? ""),
       ),
       body: SingleChildScrollView(
           child: Padding(
             padding: const EdgeInsets.all(8),
             child: Column(children: [
               row('Triggered application open', args.openedApplication.toString()),
-              row('Message ID', message.messageId ?? ""),
-              row('Sender ID', message.senderId ?? ""),
-              row('Category', message.category ?? ""),
-              row('Collapse Key', message.collapseKey ?? ""),
-              row('Content Available', message.contentAvailable.toString()),
-              row('Data', message.data.toString()),
-              row('From', message.from ?? ""),
-              row('Message ID', message.messageId ?? ""),
-              row('Sent Time', message.sentTime?.toString() ?? ""),
-              row('Thread ID', message.threadId ?? ""),
-              row('Time to Live (TTL)', message.ttl?.toString() ?? ""),
+              row('Message ID', message?.messageId ?? ""),
+              row('Sender ID', message?.senderId ?? ""),
+              row('Category', message?.category ?? ""),
+              row('Collapse Key', message?.collapseKey ?? ""),
+              row('Content Available', message?.contentAvailable.toString()),
+              row('Data', message?.data.toString()),
+              row('From', message?.from ?? ""),
+              row('Message ID', message?.messageId ?? ""),
+              row('Sent Time', message?.sentTime?.toString() ?? ""),
+              row('Thread ID', message?.threadId ?? ""),
+              row('Time to Live (TTL)', message?.ttl?.toString() ?? ""),
               if (notification != null) ...[
                 Padding(
                   padding: const EdgeInsets.only(top: 16),
