+++ UI/XPlat/flutter_template/lib/models/message_arguments.dart
@@ -1,0 +1,14 @@
+import 'package:firebase_messaging/firebase_messaging.dart';
+
+/// Message route arguments.
+class MessageArguments {
+  /// The RemoteMessage
+  final RemoteMessage message;
+
+  /// Whether this message caused the application to open.
+  final bool openedApplication;
+
+  // ignore: public_member_api_docs
+  MessageArguments(this.message, this.openedApplication)
+      : assert(message != null);
+}
